add profile pic to profile, as well as remove some code duplication
