+++ music-quiz/src/views/searchUserResultsView.js
@@ -1,35 +1,9 @@
 import { Button, Image, Row, Alert } from "react-bootstrap";
-import p1 from "../images/profile_pic_1.png";
-import p2 from "../images/profile_pic_2.png";
-import p3 from "../images/profile_pic_3.png";
-import p4 from "../images/profile_pic_4.png";
+
+import { getImgSrc } from "../settings/profileSettings";
 
 export default function SearchUserResults(props) {
-  function getProfileNumber(name) {
-    let total = 0;
-    for (let i = 0; i < name.length; i++) {
-      total += name.charCodeAt(i);
-    }
-    return (total % 4) + 1;
-  }
-
-  let rand_img_nr = getProfileNumber(props.searchResult.username);
-  let img = null;
-  switch (rand_img_nr) {
-    case 1:
-      img = p1;
-      break;
-    case 2:
-      img = p2;
-      break;
-    case 3:
-      img = p3;
-      break;
-    case 4:
-      img = p4;
-      break;
-  }
-
+  let img = getImgSrc(props.searchResult.username);
   return (
     <div>
       <Image src={img} width="70" height="70" className="me-3" />{" "}
